[Refactor] : Separate middle transport info card scale
Git : #38
