+++ presentation/home/views/lib/eb_home.dart
@@ -28,21 +28,21 @@ part 'sources/calendar_view/home_calendar_header.dart';
 part 'sources/top_schedule_info/top_schedule_info.dart';
 part 'sources/top_schedule_info/none_top_schedule_info.dart';
 part 'sources/top_schedule_info/today_top_schedule_info.dart';
-part 'sources/middle_transport/middle_transport_card_form.dart';
+part 'sources/middle_transport/cards/middle_transport_card_form.dart';
 part 'sources/home_content/schedule_addbutton.dart';
 part 'sources/bottom_list_view/home_bottom_schedule_card.dart';
 part 'sources/bottom_list_view/home_bottom_list_view.dart';
 part 'sources/home_content/home_appbar.dart';
 part 'sources/home_content/home_content.dart';
-part 'sources/middle_transport/add/middle_transport_add.dart';
+part 'sources/middle_transport/cards/add/middle_transport_add.dart';
 part 'sources/middle_transport/middle_transport_view.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info/middle_transport_info_view.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info/horizontal_scroll_card_view.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info_card/middle_transport_info_card.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info_card/left_dispatch/left_dispatch_column.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info_card/left_dispatch/left_dispatch_column_content.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info_card/left_dispatch/left_dispatch_popup_button.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info_card/left_dispatch/transport_popup_menuitem.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info_card/right_dispatch_column.dart';
-part 'sources/middle_transport/transport_info/middle_transport_info/middle_transport_info_card/reload_trasnport_info_button.dart';
-part 'sources/middle_transport/transport_info/middle_transport_imminent.dart';
+part 'sources/middle_transport/cards/info/transport_info/middle_transport_info_view.dart';
+part 'sources/middle_transport/cards/info/transport_info/middle_transport_info_card_scale.dart';
+part 'sources/middle_transport/cards/info/transport_info/middle_transport_info_card.dart';
+part 'sources/middle_transport/cards/info/transport_info/left_dispatch/left_dispatch_column.dart';
+part 'sources/middle_transport/cards/info/transport_info/left_dispatch/left_dispatch_column_content.dart';
+part 'sources/middle_transport/cards/info/transport_info/left_dispatch/left_dispatch_popup_button.dart';
+part 'sources/middle_transport/cards/info/transport_info/left_dispatch/transport_popup_menuitem.dart';
+part 'sources/middle_transport/cards/info/transport_info/right_dispatch/right_dispatch_column.dart';
+part 'sources/middle_transport/cards/info/transport_info/right_dispatch/reload_trasnport_info_button.dart';
+part 'sources/middle_transport/cards/info/imminent/middle_transport_imminent.dart';
