+++ presentation/home/views/lib/sources/middle_transport/cards/add/middle_transport_add.dart
@@ -1,4 +1,4 @@
-part of '../../../eb_home.dart';
+part of '../../../../eb_home.dart';
 
 final class MiddleTransportAddRouteCard extends StatelessWidget {
   final SchedulePath schedulePath;
