+++ presentation/home/views/lib/sources/middle_transport/cards/info/transport_info/left_dispatch/left_dispatch_column_content.dart
