+++ presentation/home/views/lib/sources/middle_transport/cards/info/transport_info/right_dispatch/reload_trasnport_info_button.dart
@@ -1,4 +1,4 @@
-part of '../../../../../eb_home.dart';
+part of '../../../../../../eb_home.dart';
 
 final class _ReloadIconTransportInfoCard extends StatefulWidget {
   final RealTimeInfo? realTimeInfo;
