+++ presentation/home/views/lib/sources/middle_transport/cards/info/transport_info/right_dispatch/right_dispatch_column.dart
@@ -1,4 +1,4 @@
-part of '../../../../../eb_home.dart';
+part of '../../../../../../eb_home.dart';
 
 final class _RightDisPatchColumn extends StatelessWidget {
   final int index;
