+++ presentation/home/views/lib/sources/middle_transport/cards/middle_transport_card_form.dart
@@ -1,4 +1,4 @@
-part of '../../eb_home.dart';
+part of '../../../eb_home.dart';
 
 final class MiddleTransportCardForm extends StatelessWidget {
   final double horizontalPadding;
